+++ neuron/src/lib/Neuron/Zettelkasten/Zettel/Parser.hs
@@ -4,6 +4,7 @@
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TupleSections #-}
+{-# LANGUAGE ViewPatterns #-}
 {-# LANGUAGE NoImplicitPrelude #-}
 
 module Neuron.Zettelkasten.Zettel.Parser where
@@ -11,7 +12,8 @@ module Neuron.Zettelkasten.Zettel.Parser where
 import Data.List (nub)
 import qualified Data.Map.Strict as Map
 import Data.Some (Some (..))
-import Data.TagTree (Tag)
+import Data.TagTree (Tag, unTagPattern)
+import qualified Data.Text as T
 import Neuron.Reader.Type (ZettelFormat, ZettelReader)
 import Neuron.Zettelkasten.ID (ZettelID (zettelIDRaw))
 import Neuron.Zettelkasten.Query.Parser (parseQueryLink)
@@ -22,6 +24,7 @@ import Neuron.Zettelkasten.Zettel
   )
 import qualified Neuron.Zettelkasten.Zettel.Meta as Meta
 import Relude
+import Text.Pandoc.Definition (Block (Plain), Inline (Code, Str), nullAttr)
 import qualified Text.Pandoc.Util as P
 
 parseZettel ::
@@ -62,14 +65,18 @@ parseZettel format zreader fn zid s = do
       case parseQueryLink uri of
         Nothing -> Nothing
         Just someQ ->
-          Just (someQ, bool mempty ctx $ isSimpleQuery someQ)
-    isSimpleQuery = \case
+          Just (someQ, convertCtx ctx someQ)
+    convertCtx :: [Block] -> Some ZettelQuery -> [Block]
+    convertCtx ctx = \case
       Some (ZettelQuery_ZettelByID _ _) ->
-        True
+        ctx
+      Some (ZettelQuery_ZettelsByTag (nonEmpty -> Nothing) _ _) ->
+        one $ Plain [Str "All zettels query"]
+      Some (ZettelQuery_ZettelsByTag (nonEmpty -> Just pats) _ _) ->
+        let tagsStr = T.intercalate ", " $ toText . unTagPattern <$> toList pats
+         in one $ Plain [Str "Linking by tag: ", Code nullAttr tagsStr]
       _ ->
-        -- Don't want to expand automatic link queries in backlinks context;
-        -- discard them.
-        False
+        mempty
 
 -- | Like `parseZettel` but operates on multiple files.
 parseZettels ::